[Feat] #WEEK3 - View VC 분리
