+++ DaangnMarket/DaangnMarket/Present/Login/ViewController/LoginViewController.swift
@@ -7,26 +7,25 @@
 
 import UIKit
 
-import Then
 import SnapKit
+import Then
 
 final class LoginViewController: UIViewController {
     
     //MARK: - UI Components
     
-    private let titleLabel = UILabel()
-    private var idTextField = UITextField()
-    private var passwordTextField = UITextField()
-    private lazy var loginButton = UIButton()
+    private let loginView = LoginView()
     
     //MARK: - Life Cycle
+    
+    override func loadView() {
+        self.view = loginView
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        setStyle()
-        setHierarchy()
-        setLayout()
+        setAddTarget()
     }
     
     @objc
@@ -39,101 +38,29 @@ final class LoginViewController: UIViewController {
 //MARK: - private Extension
 
 private extension LoginViewController {
-    func setStyle() {
-        self.view.backgroundColor = .white
-        
-        titleLabel.do {
-            $0.text = "동네라서 가능한 모든것\n당근에서 가까운 이웃과 함께해요."
-            $0.textColor = .daangnblack
-            $0.textAlignment = .center
-            $0.numberOfLines = 2
-            $0.font = .pretendardBold(size: 18)
-        }
-        
-        idTextField.do {
-            $0.placeholder = "아이디"
-        }
-        
-        passwordTextField.do {
-            $0.placeholder = "비밀번호"
-            $0.isSecureTextEntry = true
-        }
-        
-        [idTextField, passwordTextField].forEach() {
-            $0.do {
-                $0.font = .pretendardSemiBold(size: 14)
-                $0.backgroundColor = .daangngrey200
-                $0.layer.cornerRadius = 3
-                $0.clearButtonMode = .whileEditing
-                $0.addPadding(left: 23)
-            }
-        }
-        
-        loginButton.do {
-            $0.backgroundColor = .daangnprimary
-            $0.setTitle("로그인하기", for: .normal)
-            $0.setTitleColor(.daangnwhite, for: .normal)
-            $0.titleLabel?.font = .pretendardBold(size: 18)
-            $0.addTarget(self, action: #selector(loginButtonDidTap), for: .touchUpInside)
-            $0.layer.cornerRadius = 6
-        }
-    }
-    
-    func setHierarchy() {
-        [titleLabel, idTextField, passwordTextField, loginButton].forEach {
-            self.view.addSubview($0)
-        }
-    }
-    
-    //TODO: - SnapKit 적용
-    func setLayout() {
-        titleLabel.snp.makeConstraints {
-            $0.top.equalToSuperview().inset(161)
-            $0.leading.trailing.equalToSuperview().inset(69)
-            $0.height.equalTo(44)
-        }
-        
-        idTextField.snp.makeConstraints {
-            $0.top.equalTo(titleLabel.snp.bottom).offset(71)
-            $0.leading.trailing.equalToSuperview().inset(20)
-            $0.height.equalTo(52)
-        }
-        
-        passwordTextField.snp.makeConstraints {
-            $0.top.equalTo(idTextField.snp.bottom).offset(7)
-            $0.leading.trailing.equalToSuperview().inset(20)
-            $0.height.equalTo(52)
-        }
-        
-        loginButton.snp.makeConstraints {
-            $0.top.equalTo(passwordTextField.snp.bottom).offset(35)
-            $0.leading.trailing.equalToSuperview().inset(20)
-            $0.height.equalTo(58)
-        }
-        
-        
-    }
-    
     func presentToWelcomeVC() {
         let welcomeViewController = WelcomeViewController()
         welcomeViewController.modalPresentationStyle = .formSheet
-        welcomeViewController.id = idTextField.text
-        welcomeViewController.setLabelText(id: idTextField.text)
+        welcomeViewController.id = loginView.idTextField.text
+        welcomeViewController.setLabelText(id: loginView.idTextField.text)
         self.present(welcomeViewController, animated: true)
     }
     
     func pushToWelcomeVC() {
         let welcomeViewController = WelcomeViewController()
-        welcomeViewController.id = idTextField.text
-        welcomeViewController.setLabelText(id: idTextField.text)
+        welcomeViewController.id = loginView.idTextField.text
+        welcomeViewController.setLabelText(id: loginView.idTextField.text)
         self.navigationController?.pushViewController(welcomeViewController, animated: true)
     }
     
+    func setAddTarget() {
+        loginView.loginButton.addTarget(self, action: #selector(loginButtonDidTap), for: .touchUpInside)
+    }
 }
 
 extension LoginViewController {
     func resetTextField() {
-        idTextField.text = ""
-        passwordTextField.text = ""
+        loginView.idTextField.text = ""
+        loginView.passwordTextField.text = ""
     }
 }
